Move auth proxy strategy out of middleware
