+++ server/auth-strategies/register-auth-strategies.js
@@ -1,3 +1,4 @@
+const authProxy = require('./auth-proxy');
 const basic = require('./basic');
 const google = require('./google');
 const jwtServiceToken = require('./jwt-service-token');
@@ -9,6 +10,7 @@ const saml = require('./saml');
  * @param {object} config
  */
 function registerAuthStrategies(config) {
+  authProxy(config);
   basic(config);
   google(config);
   jwtServiceToken(config);
